Soothe phpstan: Remove argument validation from PeclDispatcher
The type hinting in the intercept method of the PeclDispatcher already ensures that the arguments are the correct type. Removing the explicit checks for is_object, is_string, and is_array simplifies the code. Additionally, a variable type declaration has been added to AspectTest in the test module. A new directory tests/AspectFake is also excluded from phpstan checks.